Extract queued send preparation services
